Rename IndirectBayes files to BayesFitting
