+++ qt/scientific_interfaces/Inelastic/BayesFitting/BayesFitting.cpp
@@ -4,7 +4,7 @@
 //   NScD Oak Ridge National Laboratory, European Spallation Source,
 //   Institut Laue - Langevin & CSNS, Institute of High Energy Physics, CAS
 // SPDX - License - Identifier: GPL - 3.0 +
-#include "IndirectBayes.h"
+#include "BayesFitting.h"
 #include "IndirectSettings.h"
 #include "Quasi.h"
 #include "ResNorm.h"
