+++ qt/scientific_interfaces/Inelastic/BayesFitting/BayesFitting.ui
