+++ qt/scientific_interfaces/Inelastic/BayesFitting/BayesFittingTab.cpp
@@ -4,7 +4,7 @@
 //   NScD Oak Ridge National Laboratory, European Spallation Source,
 //   Institut Laue - Langevin & CSNS, Institute of High Energy Physics, CAS
 // SPDX - License - Identifier: GPL - 3.0 +
-#include "IndirectBayesTab.h"
+#include "BayesFittingTab.h"
 
 namespace MantidQt::CustomInterfaces {
