+++ qt/scientific_interfaces/Inelastic/BayesFitting/BayesFittingTab.h
