+++ qt/scientific_interfaces/Inelastic/CMakeLists.txt
@@ -43,8 +43,8 @@ set(SRC_FILES
     Analysis/FunctionBrowser/SingleFunctionTemplateModel.cpp
     Analysis/FunctionBrowser/SingleFunctionTemplateBrowser.cpp
     Analysis/FunctionBrowser/SingleFunctionTemplatePresenter.cpp
-    BayesFitting/IndirectBayes.cpp
-    BayesFitting/IndirectBayesTab.cpp
+    BayesFitting/BayesFitting.cpp
+    BayesFitting/BayesFittingTab.cpp
     BayesFitting/Quasi.cpp
     BayesFitting/ResNorm.cpp
     BayesFitting/Stretch.cpp
@@ -100,8 +100,8 @@ set(MOC_FILES
     Analysis/FunctionBrowser/IqtTemplatePresenter.h
     Analysis/FunctionBrowser/SingleFunctionTemplateBrowser.h
     Analysis/FunctionBrowser/SingleFunctionTemplatePresenter.h
-    BayesFitting/IndirectBayes.h
-    BayesFitting/IndirectBayesTab.h
+    BayesFitting/BayesFitting.h
+    BayesFitting/BayesFittingTab.h
     BayesFitting/Quasi.h
     BayesFitting/ResNorm.h
     BayesFitting/Stretch.h
@@ -158,7 +158,7 @@ set(UI_FILES
     Analysis/IndirectFitOutputOptions.ui
     Analysis/IndirectFitTab.ui
     Analysis/IndirectSpectrumSelector.ui
-    BayesFitting/IndirectBayes.ui
+    BayesFitting/BayesFitting.ui
     BayesFitting/Quasi.ui
     BayesFitting/ResNorm.ui
     BayesFitting/Stretch.ui
